renamed tree to treeView
