+++ sites/skeleton.dev/src/lib/links.ts
@@ -104,7 +104,7 @@ export const menuNavLinks: Record<string, Array<{ title: string; list: List }>> 
 				{ href: '/components/tabs', label: 'Tabs', keywords: 'select, selection, panel' },
 				{ href: '/components/tables', label: 'Tables', keywords: 'data, entry' },
 				{ href: '/components/table-of-contents', label: 'Table of Contents', keywords: 'page, results, links, navigation' },
-				{ href: '/components/trees', label: 'Trees', keywords: 'tree, node, ', badge: 'Beta' }
+				{ href: '/components/tree-views', label: 'Tree Views', keywords: 'tree, view, node', badge: 'Beta' }
 			]
 		}
 	],
